Add files via upload
I began this project by using the print() function to display a greeting and provide instructions to the player. I then created a score variable and set its initial value to 0, which will track the number of questions the player answers correctly.
Next, I displayed a question along with multiple-choice answers to the player, allowing them to type their answer. The player's input is then stored after applying the .strip() function to remove any leading or trailing spaces, and the .lower() function to convert the input to lowercase letters.